Organized the project hierarchy
